Make it possible to paginate queries other than ORM/Query
